[gen] Put Unicode Version into reader/writer common modules
