+++ gen/src/writer/ucd/bidi.rs
@@ -15,14 +15,15 @@ use std::fmt::Display;
 use std::path::Path;
 
 use reader::ucd::unicode_data::UNICODE_DATA;
+use reader::ucd::readme::UNICODE_VERSION;
 
 use writer::utils::tables::{ToRangeCharSet, ToRangeCharTable};
-use writer::ucd::unicode_version;
+use writer::common::unicode_version;
 use writer::utils::write;
 
 
 pub fn generate(dir: &Path) {
-    unicode_version::emit(&dir);
+    unicode_version::emit(&dir, &UNICODE_VERSION);
     bidi_class_emit(&dir);
     bidi_mirrored_emit(&dir);
 }
